reorganization of the common.configuration package.
