+++ src/main/java/org/codedivoire/dembesi/common/configuration/security/AccountLogoutSuccessHandler.java
@@ -1,5 +1,6 @@
-package org.codedivoire.dembesi.common.configuration;
+package org.codedivoire.dembesi.common.configuration.security;
 
+import org.codedivoire.dembesi.common.configuration.UtilsHandler;
 import org.codedivoire.dembesi.usermanagement.entity.Profile;
 import org.codedivoire.dembesi.usermanagement.repository.ProfileRepository;
 import org.slf4j.Logger;
@@ -8,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
 import org.springframework.stereotype.Component;
 
 import javax.servlet.http.HttpServletRequest;
@@ -19,36 +20,37 @@ import java.util.Collection;
  * @author  Christian Amani on 24/08/2018.
  */
 @Component
-public class AccountAuthenticationSuccessHandler implements AuthenticationSuccessHandler,UtilsHandler {
+public class AccountLogoutSuccessHandler implements LogoutSuccessHandler, UtilsHandler {
 
-    private final Logger LOG = LoggerFactory.getLogger(AccountAuthenticationSuccessHandler.class);
+    private final Logger LOG = LoggerFactory.getLogger(AccountLogoutSuccessHandler.class);
 
     private final ProfileRepository profileRepository;
 
     @Autowired
-    public AccountAuthenticationSuccessHandler(ProfileRepository profileRepository) {
+    public AccountLogoutSuccessHandler(ProfileRepository profileRepository) {
         this.profileRepository = profileRepository;
     }
 
+
     @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
-        LOG.debug("Début du Process 'onAuthenticationSuccess'");
-        LOG.info("Succès de l'authentification");
+    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
+        LOG.debug("Début du Process 'onLogoutSuccess'");
+        LOG.info("Succès de la déconnexion");
         LOG.info("Remote IP : "+request.getRemoteAddr());
         Profile profile = (Profile) authentication.getPrincipal();
         if(profile != null) {
             LOG.info("Nom : "+profile.getUsername());
             String authorities = getAuthorities(profile);
             LOG.info("Privilege : "+authorities);
-            profile.setStatus(Profile.Status.online);
+            profile.setStatus(Profile.Status.offline);
             profileRepository.save(profile);
         }
     }
 
     @Override
-    public String getAuthorities(Profile Profile) {
+    public String getAuthorities(Profile account) {
         LOG.debug("Début du Process 'getAuthorities'");
-        Collection<? extends GrantedAuthority> grantedAuthority = Profile.getAuthorities();
+        Collection<? extends GrantedAuthority> grantedAuthority = account.getAuthorities();
         return String.join(",", AuthorityUtils.authorityListToSet(grantedAuthority));
     }
 }
