+++ src/main/java/org/codedivoire/dembesi/common/configuration/security/TokenAuthenticationFilter.java
@@ -1,11 +1,10 @@
-package org.codedivoire.dembesi.common.configuration;
+package org.codedivoire.dembesi.common.configuration.security;
 
 import com.auth0.jwt.interfaces.DecodedJWT;
 import org.codedivoire.dembesi.common.service.TokenService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.InsufficientAuthenticationException;
